Crate io_enum

io-enum

#[derive(Read, Write, Seek, BufRead)] for enums.

Usage

Add this to your Cargo.toml:

[dependencies]
io-enum = "1"

Compiler support: requires rustc 1.56+

Examples

use std::{
    fs::File,
    io::{self, Write},
    path::Path,
};

use io_enum::*;

#[derive(Read, Write, Seek, BufRead)]
enum Either<A, B> {
    A(A),
    B(B),
}

fn func(path: Option<&Path>) -> impl Write {
    if let Some(path) = path {
        Either::A(File::open(path).unwrap())
    } else {
        Either::B(io::stdout())
    }
}

See auto_enums crate for how to automate patterns like this.

Supported traits

Read - example | generated code

BufRead - example | generated code

Write - example | generated code

Seek - example | generated code

Related Projects

auto_enums: A library for to allow multiple return types by automatically generated enum.

derive_utils: A procedural macro helper for easily writing derives macros for enums.

iter-enum: #[derive(Iterator, DoubleEndedIterator, ExactSizeIterator, FusedIterator, Extend)] for enums.

License

Licensed under either of Apache License, Version 2.0 or MIT license at your option.

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Derive Macros

BufRead

Read

Seek

Write